+++ hooks/remove-dependencies-gradle-hook.js
@@ -7,6 +7,8 @@ module.exports = (context) => {
 
   const BLOCK_START = `        // DEPENDENCIES ADDED BY IMGLY - BLOCK START`;
   const BLOCK_END = `        // DEPENDENCIES ADDED BY IMGLY - BLOCK END`;
+  const VERSION_BLOCK_START = "// VERSION CHANGED BY IMGLY - START - ";
+  const VERSION_BLOCK_END = "// VERSION CHANGED BY IMGLY - END -";
 
   return new Promise((resolve, reject) => {
     const platformRoot = path.join(
@@ -30,7 +32,13 @@ module.exports = (context) => {
             fileContents.lastIndexOf(BLOCK_END) + BLOCK_END.length
           );
           fileContents = fileContents.replace(toRemove, "");
-
+          
+          
+          const regex = /VERSION CHANGED BY IMGLY - START - '([0-9]*).([0-9]*).([0-9]*)'/gm;
+          const versionMatch = regex.exec(fileContents);
+          const version = versionMatch[0].replace("VERSION CHANGED BY IMGLY - START - ", "");
+          const kotlinVersion = fileContents.substring(fileContents.indexOf(VERSION_BLOCK_START), fileContents.indexOf(VERSION_BLOCK_END) + VERSION_BLOCK_END.length);
+          fileContents = fileContents.replace(kotlinVersion, `ext.kotlin_version = ${version}`);
           fs.writeFileSync(file, fileContents, "utf8");
           console.log("remove imgly dependencies from " + file);
           callback();